# backend/docs/upload_api_swagger.json
{
    "openapi" : "3.0.3",
    "info" : {
        "title" : "Upload API",
...
        "version" : "1.0.0"
    },
    "security" : [
        {
            "BearerAuth" : []
        }
    ],
    "paths" : {
...
        "/files" : {
            "post" : {
                "tags" : [
                    "File"
                ],
                "security" : [
                    {
                        "BearerAuth" : []
                    }
                ],
                "summary" : "Create file",
                "description" : "Creates a new file and returns the identifier on success. The maximum allowed file size is set to 500Mb.",
                "requestBody" : {
                    "content" : {
                        "multipart/form-data" : {
                            "schema" : {
                                "type" : "object",
                                "properties" : {
                                    "fileName" : {
                                        "type" : "string",
                                        "format" : "binary"
                                    }
                                },
                                "required" : [
                                    "fileName"
                                ]
                            }
                        }
                    }
                },
                "responses" : {
                    "200" : {
                        "description" : "Ok",
                        "content" : {
                            "application/json" : {
                                "schema" : {
                                    "$ref" : "#/components/schemas/UploadFileResponse"
                                }
                            }
                        }
                    },
                    "400" : {
                        "description" : "Bad Request",
                        "content" : {
                            "application/json" : {
                                "schema" : {
                                    "$ref" : "#/components/schemas/BadRequest"
                                }
                            }
                        }
                    },
                    "401" : {
                        "description" : "Unauthorized",
                        "content" : {
                            "application/json" : {
                                "schema" : {
                                    "$ref" : "#/components/schemas/Unauthorized"
                                }
                            }
                        }
                    },
                    "403" : {
                        "description" : "Forbidden",
                        "content" : {
                            "application/json" : {
                                "schema" : {
                                    "$ref" : "#/components/schemas/Forbidden"
                                }
                            }
                        }
                    },
                    "500" : {
                        "description" : "Internal server error",
                        "content" : {
                            "application/json" : {
                                "schema" : {
                                    "$ref" : "#/components/schemas/InternalServerError"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/files/{UUID}" : {
            "post" : {
                "tags" : [
                    "File"
                ],
                "security" : [
                    {
                        "BearerAuth" : []
                    }
                ],
                "summary" : "Replace existing file",
                "description" : "Replaces an existing file. This is only possible when the linked RepoItem has the 'draft' status or if the file has not yet been linked.",
                "requestBody" : {
                    "content" : {
                        "multipart/form-data" : {
                            "schema" : {
                                "type" : "object",
                                "properties" : {
                                    "fileName" : {
                                        "type" : "string",
                                        "format" : "binary"
                                    }
                                },
                                "required" : [
                                    "fileName"
                                ]
                            }
                        }
                    }
                },
                "parameters" : [
                    {
                        "in" : "path",
                        "name" : "UUID",
                        "schema" : {
                            "type" : "string"
                        },
                        "required" : true,
                        "description" : "Uuid of the file to replace"
                    }
                ],
                "responses" : {
                    "201" : {
                        "description" : "Created",
                        "content" : {
                            "application/json" : {
                                "schema" : {
                                    "$ref" : "#/components/schemas/UploadFileResponse"
                                }
                            }
                        }
                    },
                    "400" : {
                        "description" : "Bad Request",
                        "content" : {
                            "application/json" : {
                                "schema" : {
                                    "$ref" : "#/components/schemas/BadRequest"
                                }
                            }
                        }
                    },
                    "401" : {
                        "description" : "Unauthorized",
                        "content" : {
                            "application/json" : {
                                "schema" : {
                                    "$ref" : "#/components/schemas/Unauthorized"
                                }
                            }
                        }
                    },
                    "403" : {
                        "description" : "Forbidden",
                        "content" : {
                            "application/json" : {
                                "schema" : {
                                    "$ref" : "#/components/schemas/Forbidden"
                                }
                            }
                        }
                    },
                    "500" : {
                        "description" : "Internal server error",
                        "content" : {
                            "application/json" : {
                                "schema" : {
                                    "$ref" : "#/components/schemas/InternalServerError"
                                }
                            }
                        }
                    }
                }
            }
        }
    },
    "components" : {
        "securitySchemes" : {
            "BearerAuth" : {
                "type" : "http",
                "scheme" : "bearer"
            }
        },
        "schemas" : {
...
            "UploadFileResponse" : {
                "type" : "object",
                "properties" : {
                    "meta" : {
                        "type" : "object",
                        "properties" : {
                            "id" : {
                                "type" : "string",
                                "example" : "e0cf09e6-680b-4423-bdff-1e1427e9f2e6"
                            }
                        }
                    }
                }
            },
            "BadRequest" : {
                "type" : "object",
                "description" : "Bad Request",
                "properties" : {
                    "code" : {
                        "type" : "string",
                        "example" : "UA_BR_001"
                    },
                    "title" : {
                        "type" : "string",
                        "example" : "Bad Request"
                    },
                    "description" : {
                        "type" : "string",
                        "example" : "client_id parameter is missing or empty"
                    }
                }
            },
...
            "Forbidden" : {
                "type" : "object",
                "description" : "Forbidden",
                "properties" : {
                    "code" : {
                        "type" : "string",
                        "example" : "UA_FB_001"
                    },
                    "title" : {
                        "type" : "string",
                        "example" : "Forbidden"
                    },
                    "description" : {
                        "type" : "string",
                        "example" : "Missing permissions for the provided institute"
                    }
                }
            },
            "Unauthorized" : {
                "type" : "object",
                "description" : "Unauthorized",
                "properties" : {
                    "code" : {
                        "type" : "string",
                        "example" : "UA_UA_001"
                    },
                    "title" : {
                        "type" : "string",
                        "example" : "Unauthorized"
                    },
                    "description" : {
                        "type" : "string",
                        "example" : "Invalid client_id and client_secret combination"
                    }
                }
            },
            "InternalServerError" : {
                "type" : "object",
                "description" : "Server Error",
                "properties" : {
                    "code" : {
                        "type" : "string",
                        "example" : "GA_ISE_001"
                    },
                    "title" : {
                        "type" : "string",
                        "example" : "Internal server error"
                    },
                    "description" : {
                        "type" : "string",
                        "example" : "An unexpected error has occurred"
                    }
                }
            }
        }
    }
}